pyproject.toml
[build-system]
requires = ["hatchling", "hatch-js", "jupyterlab>=4,<5"]
build-backend="hatchling.build"

[project]
name = "perspective-parquet"
authors = [{name = "the perspective-parquet authors", email = "t.paine154@gmail.com"}]
description = "Parquet viewer for perspective in JupyterLab"
readme = "README.md"
license = { text = "Apache-2.0" }
version = "0.1.2"
requires-python = ">=3.9"
keywords = []

classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Jupyter",
    "Framework :: Jupyter :: JupyterLab",
    "Programming Language :: Python",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "perspective-python>=3,<4",
]

[project.optional-dependencies]
develop = [
    "build",
    "bump-my-version",
    "check-manifest",
    "codespell>=2.4,<2.5",
    "hatch-js",
    "hatchling",
    "mdformat>=0.7.22,<0.8",
    "mdformat-tables>=1",
    "jupyterlab>=4,<5",
    "pytest",
    "pytest-cov",
    "ruff",
    "twine",
    "wheel",
]

[project.scripts]

[project.urls]
Repository = "https://github.com/timkpaine/perspective-parquet"
Homepage = "https://github.com/timkpaine/perspective-parquet"

[tool.bumpversion]
current_version = "0.1.2"
commit = true
tag = true
commit_args = "-s"

[[tool.bumpversion.files]]
filename = "perspective_parquet/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "js/package.json"
search = '"version": "{current_version}"'
replace = '"version": "{new_version}"'

[tool.check-manifest]
ignore = [
    ".copier-answers.yaml",
    "js/pnpm-lock.yaml",
    "Makefile",
    ".vscode/*",
    "binder/**",
    "perspective_parquet/extension/**",
    "perspective_parquet/labextension/**",
    "perspective_parquet/nbextension/**",
    "docs/**/*",
    "js/dist/**/*",
    "js/lib/*",
]

[tool.coverage.run]
branch = true
omit = [
    "perspective_parquet/tests/integration/",
]
[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]
ignore_errors = true
fail_under = 50

[tool.hatch.build]
artifacts = [
    "perspective_parquet/nbextension",
    "perspective_parquet/labextension",
]

[tool.hatch.build.sources]
src = "/"

[tool.hatch.build.targets.sdist]
packages = ["perspective_parquet", "js"]
exclude = [
    "/js/dist",
    "/js/node_modules",
]

[tool.hatch.build.targets.wheel]
packages = ["perspective_parquet"]
exclude = [
    "/js"
]

[tool.hatch.build.targets.wheel.shared-data]
"perspective_parquet/labextension" = "share/jupyter/labextensions/perspective_parquet"
"perspective_parquet/extension/install.json" = "share/jupyter/labextensions/perspective_parquet/install.json"
"perspective_parquet/extension/perspective_parquet.json" = "etc/jupyter/jupyter_server_config.d/perspective_parquet.json"

[tool.hatch.build.hooks.hatch-js]
path = "js"
build_cmd = "build"
tool = "pnpm"
targets = [
    "perspective_parquet/nbextension/static/notebook.js",
    "perspective_parquet/labextension/package.json",
]

[tool.pytest.ini_options]
addopts = ["-vvv", "--junitxml=junit.xml"]
testpaths = "perspective_parquet/tests"

[tool.ruff]
line-length = 150

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
combine-as-imports = true
default-section = "third-party"
known-first-party = ["perspective_parquet"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
